+++ src/app/modules/auth/auth.controller.ts
@@ -5,21 +5,24 @@ import httpStatus from 'http-status-codes';
 import { sendResponse } from "../../utils/sendResponse";
 import { AuthServices } from "./auth.service";
 import AppError from "../../errorHelpers/appError";
+import { setAuthCookie } from "../../utils/setCookie";
 
 
 const credentialsLogin = catchAsync (async (req: Request, res: Response, next: NextFunction) => {
     
     const loginInfo = await AuthServices.credentialsLogin(req.body);
 
-    res.cookie("accessToken", loginInfo.accessToken, {
-        httpOnly: true,
-        secure: false
-    })
+    // res.cookie("accessToken", loginInfo.accessToken, {
+    //     httpOnly: true,
+    //     secure: false
+    // })
 
-    res.cookie("refreshToken", loginInfo.refreshToken, {
-        httpOnly: true,
-        secure: false
-    });
+    // res.cookie("refreshToken", loginInfo.refreshToken, {
+    //     httpOnly: true,
+    //     secure: false
+    // });
+
+    setAuthCookie(res, loginInfo);
 
     sendResponse(res, {
         success: true,
@@ -40,6 +43,13 @@ const getNewAccessToken = catchAsync (async (req: Request, res: Response, next: 
     }
 
     const tokenInfo = await AuthServices.getNewAccessToken(refreshToken as string)
+
+    // res.cookie("accessToken", tokenInfo.accessToken, {
+    //     httpOnly: true,
+    //     secure: false
+    // })
+
+    setAuthCookie(res, tokenInfo);
 
     sendResponse(res, {
         success: true,
