+++ src/app/modules/auth/auth.service.ts
@@ -1,12 +1,10 @@
 import AppError from "../../errorHelpers/appError";
-import { IsActive, IUser } from "../user/user.interface"
+import { IUser } from "../user/user.interface"
 import { User } from "../user/user.model";
 import httpStatus from 'http-status-codes';
 import bcryptjs from 'bcryptjs';
-import { createUserTokens } from "../../utils/userTokens";
-import { generateToken, verifyToken } from "../../utils/jwt";
-import { envVars } from "../../config/env";
-import { JwtPayload } from "jsonwebtoken";
+import { createNewAccessTokenWithRefreshToken, createUserTokens } from "../../utils/userTokens";
+
 
 
 
@@ -57,32 +55,10 @@ const credentialsLogin = async (payload: Partial<IUser>) => {
 
 
 const getNewAccessToken = async (refreshToken: string) => {
-
-    const verifiedRefreshToken = verifyToken(refreshToken, envVars.JWT_REFRESH_SECRET) as JwtPayload;
-
-    const isUserExit = await User.findOne({ email: verifiedRefreshToken.email });
-
-    if (!isUserExit) {
-        throw new AppError(httpStatus.BAD_REQUEST, "User does not exist")
-    }
-    if (isUserExit.isActive === IsActive.BLOCKED || isUserExit.isActive === IsActive.INACTIVE) {
-        throw new AppError(httpStatus.BAD_REQUEST, `User is ${isUserExit.isActive}`)
-    }
-    if (isUserExit.isDeleted) {
-        throw new AppError(httpStatus.BAD_REQUEST, "User is deleted")
-    }
-
-    const jwtPayload = {
-        userId : isUserExit._id,
-        email: isUserExit.email,
-        role: isUserExit.role
-    }
-
-    // const accessToken = jwt.sign(jwtPayload, "secret", {expiresIn: "1d"})
-    const accessToken = generateToken(jwtPayload, envVars.JWT_ACCESS_SECRET, envVars.JWT_ACCESS_EXPIRES)
+    const newAccessToken = await createNewAccessTokenWithRefreshToken(refreshToken)
 
     return {
-        accessToken,
+        accessToken : newAccessToken
     }
 }
