+++ src/app/utils/setCookie.ts
@@ -1,0 +1,22 @@
+import { Response } from "express";
+
+export interface AuthTokens {
+    accessToken?: string;
+    refreshToken?: string
+}
+
+export const setAuthCookie = (res: Response, tokenInfo: AuthTokens) => {
+    if (tokenInfo.accessToken) {
+        res.cookie("accessToken", tokenInfo.accessToken, {
+            httpOnly: true,
+            secure: false
+        })
+    }
+
+    if (tokenInfo.refreshToken) {
+        res.cookie("refreshToken", tokenInfo.refreshToken, {
+            httpOnly: true,
+            secure: false
+        });
+    }
+}
