+++ src/app/utils/userTokens.ts
@@ -1,6 +1,10 @@
+import { JwtPayload } from "jsonwebtoken";
 import { envVars } from "../config/env";
-import { IUser } from "../modules/user/user.interface";
-import { generateToken } from "./jwt";
+import { IsActive, IUser } from "../modules/user/user.interface";
+import { generateToken, verifyToken } from "./jwt";
+import { User } from "../modules/user/user.model";
+import AppError from "../errorHelpers/appError";
+import httpStatus from 'http-status-codes';
 
 export const createUserTokens = (user: Partial<IUser>) => {
     const jwtPayload = {
@@ -17,4 +21,32 @@ export const createUserTokens = (user: Partial<IUser>) => {
         accessToken,
         refreshToken
     }
+}
+
+
+export const createNewAccessTokenWithRefreshToken = async (refreshToken: string) => {
+    const verifiedRefreshToken = verifyToken(refreshToken, envVars.JWT_REFRESH_SECRET) as JwtPayload;
+
+    const isUserExit = await User.findOne({ email: verifiedRefreshToken.email });
+
+    if (!isUserExit) {
+        throw new AppError(httpStatus.BAD_REQUEST, "User does not exist")
+    }
+    if (isUserExit.isActive === IsActive.BLOCKED || isUserExit.isActive === IsActive.INACTIVE) {
+        throw new AppError(httpStatus.BAD_REQUEST, `User is ${isUserExit.isActive}`)
+    }
+    if (isUserExit.isDeleted) {
+        throw new AppError(httpStatus.BAD_REQUEST, "User is deleted")
+    }
+
+    const jwtPayload = {
+        userId: isUserExit._id,
+        email: isUserExit.email,
+        role: isUserExit.role
+    }
+
+    // const accessToken = jwt.sign(jwtPayload, "secret", {expiresIn: "1d"})
+    const accessToken = generateToken(jwtPayload, envVars.JWT_ACCESS_SECRET, envVars.JWT_ACCESS_EXPIRES)
+
+    return accessToken;
 }
